+++ NEWS.md
@@ -12,3 +12,6 @@
 * Added FCS colors (v0.0.0.9007)
 * Added DII & DIII Logos
 * Added input to `valid_team_names()` to allow filtering by division (FBS, FCS, DII, DIII, or Conference logos) (v0.0.0.9008)
+* Added the function `cfb_team_tiers()` that build an NFL team tiers ggplot, thanks to [Timo Riske](https://twitter.com/PFF_Moo) for the suggestion.
+* Fixed a bug in `geom_median_lines()` and `geom_mean_lines()` that caused `alpha` to not work properly.
+* Added the `geom_cfb_headshots()` geom that plots headshots for valid ESPN player IDs. (v0.0.9009)